Major Flaws In Meta Model Corrected
Corrected misinterpretation of the "upper bound" and "lower bound"
properties of class attributes. These properties don't constrain the
allowed range of the value itself but the count of allowed values for
that attribute (e.g. an upper bound of seven allows the user to enter an
array of up to seven elements - values, that is, for that attribute).